EventPassContractReveralButtonClient : add NextIntl provider to fix whitescreen and translations issues. Also moved reveal to client side for thirdweb purpose
