+++ libs/features/back-office/events/src/lib/actions/deployCollectionWrapper.ts
@@ -1,4 +1,5 @@
 'use client';
+
 import env from '@env/client';
 import { EventPass } from '@features/back-office/events-types';
 import { EventPassNftContractType_Enum } from '@gql/shared/types';
