+++ libs/features/back-office/events/src/lib/molecules/EventPassNftFilesTable/examples.tsx
@@ -13,7 +13,6 @@ import * as deleteFile from '../../actions/deleteEventPassFile';
 import * as deploy from '../../actions/deployCollectionWrapper';
 import * as getPass from '../../actions/getEventPassNftFiles';
 import * as renameFiles from '../../actions/renameEventPassNftFiles';
-import * as reveal from '../../actions/revealDelayedContract';
 import { EventPassNftFilesTableProps } from './EventPassNftFilesTable';
 
 export const eventPassNftFiles = [
@@ -151,8 +150,6 @@ export function eventPassNftFilesTableMocks() {
   mockRename.mockReturnValue(Promise.resolve());
   const mockDeploy = createMock(deploy, 'deployCollectionWrapper');
   mockDeploy.mockReturnValue(Promise.resolve());
-  const mockReveal = createMock(reveal, 'revealDelayedContract');
-  mockReveal.mockReturnValue(Promise.resolve());
 
   return [
     mock,
@@ -163,7 +160,6 @@ export function eventPassNftFilesTableMocks() {
     mockCheckPass,
     mockRename,
     mockDeploy,
-    mockReveal,
     ...i18nUiTablesServerMocks(),
   ];
 }
